Revert "Creating delete post method"
